refactor: merge static/file.go into static/middleware.go
file.go contained only two methods (openFile, openIndexFile) that
are exclusively called by the Middleware struct. No reason to split
64 lines of tightly coupled logic into a separate file.
